// src/api/generated/model/adminCreateSpaceRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Fiber OpenAPI
 * API documentation generated by fiber-oapi
 * OpenAPI spec version: 1.0.0
 */
import type { AdminCreateSpaceRequestType } from './adminCreateSpaceRequestType';

export interface AdminCreateSpaceRequest {
  icon?: string;
  icon_color?: string;
  /**
   * @minLength 1
   * @maxLength 100
   */
  name: string;
  owner_id?: string;
  type: AdminCreateSpaceRequestType;
}