// src/api/generated/model/adminAddSpaceUserPermissionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Fiber OpenAPI
 * API documentation generated by fiber-oapi
 * OpenAPI spec version: 1.0.0
 */
import type { AdminAddSpaceUserPermissionRequestRole } from './adminAddSpaceUserPermissionRequestRole';

export interface AdminAddSpaceUserPermissionRequest {
  role: AdminAddSpaceUserPermissionRequestRole;
  user_id: string;
}