// src/api/generated/model/adminAddSpaceGroupPermissionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Fiber OpenAPI
 * API documentation generated by fiber-oapi
 * OpenAPI spec version: 1.0.0
 */
import type { AdminAddSpaceGroupPermissionRequestRole } from './adminAddSpaceGroupPermissionRequestRole';

export interface AdminAddSpaceGroupPermissionRequest {
  group_id: string;
  role: AdminAddSpaceGroupPermissionRequestRole;
}