// src/api/generated/model/inviteUserRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Fiber OpenAPI
 * API documentation generated by fiber-oapi
 * OpenAPI spec version: 1.0.0
 */
import type { InviteUserRequestRole } from './inviteUserRequestRole';

export interface InviteUserRequest {
  email: string;
  role: InviteUserRequestRole;
}