// src/api/generated/model/createGroupRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Fiber OpenAPI
 * API documentation generated by fiber-oapi
 * OpenAPI spec version: 1.0.0
 */
import type { CreateGroupRequestRole } from './createGroupRequestRole';

export interface CreateGroupRequest {
  /** @maxLength 500 */
  description?: string;
  /**
   * @minLength 1
   * @maxLength 100
   */
  name: string;
  role: CreateGroupRequestRole;
}